// src/app/models/proveedor.model.ts
export class ProveedorModel {
    provId?:            number;
    provRuc?:           string;
    provNomComercial?:  string;
    provRazon?:         string;
    provDireccion?:     string;
    provTelefono?:      string;
    ciudadId?:          number;
    ciudadDescripcion?: string;
    estadoId?:          number;
    estadoDescripcion?: string;
    fechaHoraReg?:      Date;
    usuIdReg?:          number;
    usuRegName?:        string;
}